necc information migrated over to sendmessages component
