+++ client/src/components/DisplayMessages/assets/css/usermessages.css
@@ -1,13 +1,3 @@
-.message-button {
-    width: 100%;
-    border-bottom-left-radius: 5px;
-    border-bottom-right-radius: 5px;
-    font-family: var(--chewy);
-    background-color: var(--yellow);
-    color: var(--PPshade1);
-    font-size: 30px;
-}
-
 .message-div {
     padding: 1%;
     display: flex;
@@ -17,18 +7,6 @@
     border-left: #000 solid 2px;
     border-right: #000 solid 2px;
     background: linear-gradient(180deg, rgba(217,210,215,1) 7%, rgba(255,247,253,1) 100%);
-}
-
-.message-form {
-    width: 100%;
-}
-
-.message-input {
-    height: 40px;
-    width: 100%;
-    border-left: #000 solid 2px;
-    border-right: #000 solid 2px;
-    padding: 1%;
 }
 
 #user-heading {
