+++ client/src/components/DisplayMessages/index.js
@@ -8,7 +8,6 @@ const DisplayMessages = () => {
 
     const currentGroup = localStorage.getItem('currentGroupChat');
     const currentUser = localStorage.getItem('currentUser');
-    const [formState, setFormState] = useState({ message: '' });
 
     const { loading, data } = useQuery(QUERY_GROUP,
         {
@@ -38,30 +37,6 @@ const DisplayMessages = () => {
 
     });
 
-
-
-    const handleChange = (e) => {
-        const { name, value } = e.target;
-
-        setFormState({
-            ...formState,
-            [name]: value,
-        });
-
-        console.log(formState);
-    };
-
-    const handleFormSubmit = async (e) => {
-        e.preventDefault();
-        e.stopPropagation();
-
-        setFormState({
-            message: ''
-        });
-    };
-
-
-
     return (
         <>
             <div className='message-div'>
@@ -69,11 +44,6 @@ const DisplayMessages = () => {
                 <h2 id="user-heading">Logged in as: {currentUser}</h2>
                     {loading ? 'Loading..' : handleMessages}
             </div>
-            {/* TODO: Add button send message functionality */}
-            <form className="message-form" onSubmit={handleFormSubmit}>
-                <input id="message" className="message-input" placeholder=". . . my super sick message" name="message" type="text" onChange={handleChange} />
-                <button id="send" className='message-button' type="submit" >Send Message</button>
-            </form>
         </>
     )
 }
